http: Implement get_block_height
Switch from HTTP -> RPC-wire-protocol calls to HTTP -> RPC internal
calls. This should be squashed to remove the earlier changes that this
was based on when rebasing to `master`

The RPC wire protocol portion relied on the RPC endpoint being exposed
and tied post-HTTP processing back to the RPC protocol parsing.

Use a mutex around the Electrum/Rpc struct to allow the Axum side to
reach the underlying data. The mutex is currently overly-broad to mimize
changes required.

Future work implementing the HTTP API would need to `pub` expose more of
the `Rpc` methods or splitting the RPC wire protocol handling from the
calls to the underlying Tracker/Cache.